refactor catalog updater
