+++ app/catalog/catalog_operations.py
@@ -1,0 +1,85 @@
+"""
+This is the module of catalog updater.
+
+It contains functions for create and update catalog of bicycles from canyon web page.
+"""
+
+from typing import List, Tuple
+
+import httpx
+from lxml import etree
+
+from app import storage
+from app.models import Bike
+from app.settings import app_settings
+
+
+def get_canyon_catalog_html() -> etree._Element:  # noqa: WPS437
+    """Get HTML from the canyon catalog web page. Return HTML."""
+    query_params = {
+        'cgid': 'orderable-bikes',
+        'prefn1': 'isInStock',
+        'prefv1': 'In-stock',
+        'start': '0',
+        'sz': '300',
+        'searchredirect': 'false',
+        'pn': '1',
+        'format': 'ajax',
+    }
+
+    catalog_response = httpx.get(
+        'https://www.canyon.com/on/demandware.store/Sites-RoW-Site/en_CZ/Search-IncludeProductGrid',
+        timeout=app_settings.timeout,
+        params=query_params,
+    )
+    html_source: str = catalog_response.text
+
+    return etree.HTML(html_source)
+
+
+def normalize_bike_id(bike_title: str) -> str:
+    """Bring the id to the same view: lowercase, underscore instead of whitespace. Return bike id normalized."""
+    return bike_title.replace(' ', '_').lower()
+
+
+def parse_canyon_catalog(html_tree: etree._Element) -> List[Bike]:  # noqa: WPS437, WPS210
+    """Make the list of bike elements from HTML. Return list of bikes in elements."""
+    output: List[Bike] = []
+
+    html_bike_list = html_tree.cssselect('.productGrid__listItem')
+
+    for list_item in html_bike_list:
+        bike_name_element: etree.Element = list_item.cssselect('.productTile__link')[0]
+
+        bike_title_list = bike_name_element.get('title').split(' ')
+        if bike_name_element.get('title').startswith('Grand Canyon'):
+            bike_family = f'{bike_title_list[0]} {bike_title_list[1]}'
+            bike_model = ' '.join(bike_title_list[2:])
+        else:
+            bike_family = bike_title_list[0]
+            bike_model = ' '.join(bike_title_list[1:])
+
+        bike_item: Bike = Bike(
+            id=normalize_bike_id(bike_name_element.get('title')),
+            title=bike_name_element.get('title'),
+            link=bike_name_element.get('href'),
+            family=bike_family,
+            model=bike_model,
+        )
+        output.append(bike_item)
+
+    return output
+
+
+async def update_catalog(actual_catalog: List[Bike]) -> Tuple[int, int]:
+    """Clear the old catalog in database and insert actual. Return amount of deleted and added items."""
+    items_deleted: int = await storage.clear_catalog()
+    items_added: int = await storage.insert_actual_catalog(actual_catalog)
+
+    return items_deleted, items_added
+
+
+def get_canyon_catalog() -> List[Bike]:
+    """Get html from the web page. Return list of bikes in elements."""
+    html_tree = get_canyon_catalog_html()
+    return parse_canyon_catalog(html_tree)
