+++ pyproject.toml
@@ -38,3 +38,7 @@ filterwarnings = [
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.isort]
+line_length = 125
+profile = "black"
