+++ tests/tests_catalog_updater/test_get_canyon_catalog.py
@@ -1,9 +1,9 @@
-from app.catalog_updater import _get_canyon_catalog
+from app.catalog.catalog_operations import get_canyon_catalog
 from app.models import Bike
 
 
 def test_get_canyon_catalog():
-    res = _get_canyon_catalog()
+    res = get_canyon_catalog()
 
     assert isinstance(res, list)
     assert len(res) > 0
