+++ tests/tests_catalog_updater/test_get_canyon_catalog_html.py
@@ -1,7 +1,7 @@
-from app.catalog_updater import _get_canyon_catalog_html
+from app.catalog.catalog_operations import get_canyon_catalog_html
 
 
 def test_get_canyon_catalog_html():
-    res = _get_canyon_catalog_html()
+    res = get_canyon_catalog_html()
 
     assert res is not None
