+++ tests/tests_catalog_updater/test_get_new_available_bikes.py
@@ -1,5 +1,5 @@
+from app.catalog.available_bike_list_operations import get_new_available_bikes
 from app.models import Bike
-from app.catalog_updater import _get_new_available_bikes
 
 
 def test_get_new_available_bikes_happy_path():
@@ -36,7 +36,7 @@ def test_get_new_available_bikes_happy_path():
         ),
     ]
 
-    res = _get_new_available_bikes(old_bikes_list, new_bikes_list)
+    res = get_new_available_bikes(old_bikes_list, new_bikes_list)
 
     assert res == [
         Bike(
@@ -68,7 +68,7 @@ def test_get_new_available_bikes_empty_old_list():
         ),
     ]
 
-    res = _get_new_available_bikes(old_bikes_list, new_bikes_list)
+    res = get_new_available_bikes(old_bikes_list, new_bikes_list)
 
     assert res == new_bikes_list
 
@@ -92,7 +92,7 @@ def test_get_new_available_bikes_empty_new_list():
     ]
     new_bikes_list = []
 
-    res = _get_new_available_bikes(old_bikes_list, new_bikes_list)
+    res = get_new_available_bikes(old_bikes_list, new_bikes_list)
 
     assert res == []
 
@@ -101,6 +101,6 @@ def test_get_new_available_bikes_empty_all_lists():
     old_bikes_list = []
     new_bikes_list = []
 
-    res = _get_new_available_bikes(old_bikes_list, new_bikes_list)
+    res = get_new_available_bikes(old_bikes_list, new_bikes_list)
 
     assert res == []
