+++ tests/tests_catalog_updater/test_normalize_bike_id.py
@@ -1,6 +1,6 @@
 import pytest
 
-from app.catalog_updater import normalize_bike_id
+from app.catalog.catalog_operations import normalize_bike_id
 
 
 @pytest.mark.parametrize("test_input,expected", [("Spectral 125 CF 9", "spectral_125_cf_9"), ("", "")])
