+++ tests/tests_catalog_updater/test_update_catalog.py
@@ -1,10 +1,10 @@
 from typing import List
 
+from app.catalog.catalog_operations import update_catalog
 from app.models import Bike
-from app.catalog_updater import _update_catalog
 
 
 async def test_update_catalog_happy_path(fixture_prefilled_catalog: List[Bike]):
-    res = await _update_catalog(fixture_prefilled_catalog)
+    res = await update_catalog(fixture_prefilled_catalog)
 
     assert res == (2, 2)
